moved SearchPage to features
