+++ src/features/SearchPage/SearchedMovies.js
@@ -1,11 +1,11 @@
 import { useSelector } from "react-redux";
-import { Main } from "../../../common/Main/Main";
-import { Section, SectionTitle } from "../../../common/Section/Section";
-import { LargeListWrapper, StyledLink } from "../../../common/Tile/styled";
-import { ListTileLarge } from "../../../common/Tile/index";
-import AnimatedPage from "../../../common/AnimatedPage";
-import Pagination from "../../../common/Pagination";
-import { incrementPage, decrementPage, goToFirstSearchPage, goToLastSearchPage, selectCurrnetPage, selectTotalPages } from "../../../common/Navigation/Search/searchSlice";
+import { Main } from "../../common/Main/Main";
+import { Section, SectionTitle } from "../../common/Section/Section";
+import { LargeListWrapper, StyledLink } from "../../common/Tile/styled";
+import { ListTileLarge } from "../../common/Tile/index";
+import AnimatedPage from "../../common/AnimatedPage";
+import Pagination from "../../common/Pagination";
+import { incrementPage, decrementPage, goToFirstSearchPage, goToLastSearchPage, selectCurrnetPage, selectTotalPages } from "../../common/Navigation/Search/searchSlice";
 
 export const SearchedMovies = ({ searchQuery, searchResults, totalResults }) => {
     const result = searchResults;
