+++ src/features/SearchPage/styled.js
